dw: fix and change
add top build host like DNS
change default home to /local
try to fix queue clean issue

+++ ihook/device_matching.sh
@@ -23,7 +23,7 @@ export TASK_SPACE=/dev/shm
 export LOCK_SPACE=/dev/shm/lock
 export SEED=$RANDOM
 [[ `echo $* | grep debug` ]] && export DEBUG=echo
-export HOME=/root
+[[ ! -d $HOME/ibuild ]] && export HOME=/local
 
 export IBUILD_ROOT=$HOME/ibuild
         [[ -z $IBUILD_ROOT ]] && export IBUILD_ROOT=$(dirname $0 | awk -F'/ibuild' {'print $1'})'/ibuild'
@@ -74,8 +74,12 @@ MATCHING()
 #     export ASSIGN_HOST_DEVICE=$(cat $IVERFY_SPACE/$HOST_DEVICE.assign | awk -F'|' {'print $3'})
 #     if [[ $ASSIGN_HOST_DEVICE = $HOST_DEVICE ]] ; then
      if [[ `grep $HOSTNAME_DEVICE $IVERFY_SPACE/$HOST_DEVICE.assign` ]] ; then
-         svn rm --force $QUEUE_SPACE/$PRIORITY_ICASE_REV
-         rm -f $IVERFY_SPACE/inode.lock/$HOST_DEVICE
+         svn rm -q $IBUILD_SVN_OPTION -m "auto: remove $PRIORITY_ICASE_REV" svn://$IBUILD_SVN_SRV/istatus/queue/icase/$PRIORITY_ICASE_REV
+         svn up -q $IBUILD_SVN_OPTION $QUEUE_SPACE
+         if [[ -f $QUEUE_SPACE/$PRIORITY_ICASE_REV ]] ; then
+             svn rm --force $QUEUE_SPACE/$PRIORITY_ICASE_REV
+             rm -f $IVERFY_SPACE/inode.lock/$HOST_DEVICE
+         fi
          EXIT
      fi
  done
